+++ kernel-rs/src/uart.rs
@@ -9,7 +9,7 @@ use core::ptr;
 /// address of one of the registers.
 #[inline]
 const fn reg(r: i32) -> *mut u8 {
-    (UART0 + r as u64) as *mut u8
+    (UART0 + r as usize) as *mut u8
 }
 
 /// the UART control registers.
